requirements.txt
ccxt==1.95.11
Django==4.1.2
djangorestframework==3.14.0
pytest==7.1.3
python-telegram-bot==13.11
requests==2.28.1